+++ processor/src/test/resources/bad_input/com/example/post_mount_or_update/AbstractModel.java
@@ -1,19 +1,16 @@
-package com.example.post_render;
+package com.example.post_mount_or_update;
 
 import react4j.Component;
 import react4j.ReactNode;
-import react4j.annotations.PostRender;
-import react4j.annotations.PreUpdate;
+import react4j.annotations.PostMountOrUpdate;
 import react4j.annotations.ReactComponent;
 
 @ReactComponent
-abstract class BasicModel
+abstract class AbstractModel
   extends Component
 {
-  @PostRender
-  void postRender()
-  {
-  }
+  @PostMountOrUpdate
+  abstract void postMountOrUpdate();
 
   @Override
   protected ReactNode render()