+++ processor/src/test/resources/bad_input/com/example/post_mount_or_update/ParametersModel.java
@@ -1,16 +1,16 @@
-package com.example.post_render;
+package com.example.post_mount_or_update;
 
 import react4j.Component;
 import react4j.ReactNode;
-import react4j.annotations.PostRender;
+import react4j.annotations.PostMountOrUpdate;
 import react4j.annotations.ReactComponent;
 
 @ReactComponent
-abstract class StaticModel
+abstract class ParametersModel
   extends Component
 {
-  @PostRender
-  static void postRender()
+  @PostMountOrUpdate
+  void postRender( int x )
   {
   }