+++ processor/src/test/resources/bad_input/com/example/post_mount_or_update/PrivateModel.java
@@ -1,18 +1,17 @@
-package com.example.post_render;
+package com.example.post_mount_or_update;
 
 import react4j.Component;
 import react4j.ReactNode;
-import react4j.annotations.PostRender;
+import react4j.annotations.PostMountOrUpdate;
 import react4j.annotations.ReactComponent;
 
 @ReactComponent
-abstract class ReturnsValueModel
+abstract class PrivateModel
   extends Component
 {
-  @PostRender
-  int postRender()
+  @PostMountOrUpdate
+  private void postMountOrUpdate()
   {
-    return 0;
   }
 
   @Override