+++ processor/src/test/resources/bad_input/com/example/post_mount_or_update/PublicModel.java
@@ -1,18 +1,16 @@
-package com.example.post_render;
+package com.example.post_mount_or_update;
 
-import java.io.IOException;
 import react4j.Component;
 import react4j.ReactNode;
-import react4j.annotations.PostRender;
+import react4j.annotations.PostMountOrUpdate;
 import react4j.annotations.ReactComponent;
 
 @ReactComponent
-abstract class ThrowsModel
+abstract class PublicModel
   extends Component
 {
-  @PostRender
-  void postRender()
-    throws IOException
+  @PostMountOrUpdate
+  public void postMountOrUpdate()
   {
   }