+++ processor/src/test/resources/bad_input/com/example/post_mount_or_update/ReturnsValueModel.java
@@ -1,17 +1,18 @@
-package com.example.post_render;
+package com.example.post_mount_or_update;
 
 import react4j.Component;
 import react4j.ReactNode;
-import react4j.annotations.PostRender;
+import react4j.annotations.PostMountOrUpdate;
 import react4j.annotations.ReactComponent;
 
 @ReactComponent
-abstract class ParametersModel
+abstract class ReturnsValueModel
   extends Component
 {
-  @PostRender
-  void postRender( int x )
+  @PostMountOrUpdate
+  int postMountOrUpdate()
   {
+    return 0;
   }
 
   @Override