+++ processor/src/test/resources/bad_input/com/example/post_mount_or_update/StaticModel.java
@@ -1,21 +1,16 @@
-package com.example.post_render;
+package com.example.post_mount_or_update;
 
 import react4j.Component;
 import react4j.ReactNode;
-import react4j.annotations.PostRender;
+import react4j.annotations.PostMountOrUpdate;
 import react4j.annotations.ReactComponent;
 
 @ReactComponent
-abstract class DuplicateModel
+abstract class StaticModel
   extends Component
 {
-  @PostRender
-  void postRender()
-  {
-  }
-
-  @PostRender
-  void postRender2()
+  @PostMountOrUpdate
+  static void postMountOrUpdate()
   {
   }