+++ processor/src/test/resources/bad_input/com/example/post_mount_or_update/ThrowsModel.java
@@ -1,16 +1,18 @@
-package com.example.post_render;
+package com.example.post_mount_or_update;
 
+import java.io.IOException;
 import react4j.Component;
 import react4j.ReactNode;
-import react4j.annotations.PostRender;
+import react4j.annotations.PostMountOrUpdate;
 import react4j.annotations.ReactComponent;
 
 @ReactComponent
-abstract class PrivateModel
+abstract class ThrowsModel
   extends Component
 {
-  @PostRender
-  private void postRender()
+  @PostMountOrUpdate
+  void postMountOrUpdate()
+    throws IOException
   {
   }