+++ processor/src/test/resources/input/com/example/post_mount_or_update/BasicModel.java
@@ -1,16 +1,16 @@
-package com.example.post_render;
+package com.example.post_mount_or_update;
 
 import react4j.Component;
 import react4j.ReactNode;
-import react4j.annotations.PostRender;
+import react4j.annotations.PostMountOrUpdate;
 import react4j.annotations.ReactComponent;
 
 @ReactComponent
-abstract class PublicModel
+abstract class BasicModel
   extends Component
 {
-  @PostRender
-  public void postRender()
+  @PostMountOrUpdate
+  void postMountOrUpdate()
   {
   }